feat: 대화 진행 여부 선택지 삭제
