+++ frontend/src/business/hooks/useAiChat/index.ts
@@ -1,2 +1,2 @@
 export { useAiChatMessage } from './useAiChatMessage';
-export { useAiChatTarot } from './useAiChatTarot';
+export { useTarotSpread } from './useTarotSpread';
